+++ src/main/java/site/siredvin/gttruesteam/TrueSteamConcepts.java
@@ -34,7 +34,7 @@ public class TrueSteamConcepts {
 
     public static Concept SteamConcept = register(
             RecipeConcept.create(
-                    "steam", 0xc9c9c9, GTRecipeTypes.STEAM_TURBINE_FUELS, 4, Collections.emptyList()));
+                    "steam", 0xc9c9c9, GTRecipeTypes.STEAM_TURBINE_FUELS, 4, List.of(MaterialFlags.GENERATE_ROD)));
 
     public static Concept CoolingConcept = register(
             CombinedConcept.create("cooling", 0x107AB0, List.of(HeatingConcept, PolarizationConcept), List.of(