+++ src/main/java/site/siredvin/gttruesteam/TrueSteamRecipeConditions.java
@@ -4,6 +4,7 @@ import com.gregtechceu.gtceu.api.GTCEuAPI;
 import com.gregtechceu.gtceu.api.recipe.condition.RecipeConditionType;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
 
+import site.siredvin.gttruesteam.recipe.condition.BeatingHuskCondition;
 import site.siredvin.gttruesteam.recipe.condition.CoatingFluidCondition;
 import site.siredvin.gttruesteam.recipe.condition.CoolingCapacityCondition;
 import site.siredvin.gttruesteam.recipe.condition.InnerRecipeTypeCondition;
@@ -13,6 +14,7 @@ public class TrueSteamRecipeConditions {
     public static RecipeConditionType<CoatingFluidCondition> COATING_FLUID;
     public static RecipeConditionType<CoolingCapacityCondition> COOLING_CAPACITY;
     public static RecipeConditionType<InnerRecipeTypeCondition> INNER_RECIPE_TYPE;
+    public static RecipeConditionType<BeatingHuskCondition> BEATING_HUSK_TYPE;
 
     public static void registerConditions(GTCEuAPI.RegisterEvent<String, RecipeConditionType<?>> event) {
         COATING_FLUID = GTRegistries.RECIPE_CONDITIONS.register("coating_fluid", //
@@ -21,5 +23,7 @@ public class TrueSteamRecipeConditions {
                 new RecipeConditionType<>(CoolingCapacityCondition::new, CoolingCapacityCondition.CODEC));
         INNER_RECIPE_TYPE = GTRegistries.RECIPE_CONDITIONS.register("inner_recipe_type", //
                 new RecipeConditionType<>(InnerRecipeTypeCondition::new, InnerRecipeTypeCondition.CODEC));
+        BEATING_HUSK_TYPE = GTRegistries.RECIPE_CONDITIONS.register("beating_husk_type", //
+                new RecipeConditionType<>(BeatingHuskCondition::new, BeatingHuskCondition.CODEC));
     }
 }