+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New infernal boiler textures (Thanks to MindBrain27)
 - Alternative boiler charging fluid based on nether trees.
 - New cooling coils from cooling cometal
+- Update for infernal boiler in form of beating husks
 
 ### Changed